Run Aider automatically if it is not already running
Currently, if you try to send a command to Aider and the session is not running, it will just raise
an error. Instead what we can do is just start the session for the user.

When I want to add a file to aider with the old flow, it would look something like this:

1. run `aidermacs-add-current-file` via the transient.
2. it tells me that aider is not running
3. run `aidermacs-run`. This switches me away from my current buffer
4. Switch back to my buffer
5. Run `aidermacs-add-current-file` again.

After this change, running `aidermacs-add-current-file` will just work the first time.