Use redirection for sending commands to aider
I changed to use redirection for drop files. However I don't know if I fully understand all the
systems at play, so please double check.

Added ./ to the start of the dropped file to distinguish files that have the same
suffix. (i.e. foo.py and test_foo.py would both drop if you run `/drop foo.py`).

Also, the ls command will list what files are in the repo but not in the chat. We should be able to
parse that and create a command to easily add new files from the list.