Don't update diff headers in font lock
Closes #50

The issue here is that diff-mode by default will try and update the line numbers in the headers on
the fly. But this leads to the font lock buffer and the actual buffer getting out of sync. For
example if it changes from

```diff
@@ -1,36 +1,3 @@
```

to

```diff
@@ -1,9 +1,0 @@
```

The two positions will now be off by 1 (because 36 was changed to 9). To fix we disable this behavior.