aidermacs--form-prompt: Always use prompt-prefix and context.
Currently, user input (user-command) must be non-empty for it to be concatenated
with prompt-prefix and context. Otherwise, only the command (e.g., /code,
/architecture...) is sent to aider, effectively doing nothing.

This is confusing when the prompt-prefix is sufficient to form the complete
prompt, such as with `aidermacs-write-unit-test`. The user must enter something
for the prompt to be formed and sent; otherwise, only `/architect` is sent.

This change simplifies the logic in `aidermacs--form-prompt`. It will always use
prompt-prefix and context. If these variables are nil, `concat` will ignore
them.